feat: add a utils module
