+++ CMakeLists.txt
@@ -5,9 +5,9 @@ set(CMAKE_CXX_STANDARD 20)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 
 # Check for BOOST_ROOT environment variable
-if(NOT DEFINED ENV{BOOST_ROOT})
+if (NOT DEFINED ENV{BOOST_ROOT})
     message(FATAL_ERROR "BOOST_ROOT environment variable is required but not set")
-endif()
+endif ()
 
 # Set Boost directory from environment
 set(BOOST_ROOT $ENV{BOOST_ROOT})
@@ -23,22 +23,16 @@ message(STATUS "OpenSSL Libraries: ${OPENSSL_LIBRARIES}")
 # find_package(OpenSSL REQUIRED COMPONENTS Crypto SSL) might help
 message(STATUS "OpenSSL Version: ${OPENSSL_VERSION}") # May or may not be set
 
-# Common source files
-set(SOURCES_COMMON
-
-)
-
 # Add source files
 set(SOURCES_VALIDATOR
         catl-validator.cpp
-        ${SOURCES_COMMON}
 )
 
 set(SOURCES_HASHER
         catl-hasher.cpp
         core-types.cpp
+        utils.cpp
         logger.cpp
-        ${SOURCES_COMMON}
 )
 
 # Create executables
@@ -46,41 +40,41 @@ add_executable(catl-validator ${SOURCES_VALIDATOR})
 add_executable(catl-hasher ${SOURCES_HASHER})
 
 # Add include directories
-foreach(target catl-validator catl-hasher)
+foreach (target catl-validator catl-hasher)
     target_include_directories(${target} PRIVATE
             ${BOOST_ROOT}
             ${OPENSSL_INCLUDE_DIR}
             ${ZLIB_INCLUDE_DIRS}
-            ${CMAKE_CURRENT_SOURCE_DIR}  # This ensures the targets can find Logger.h
+            ${CMAKE_CURRENT_SOURCE_DIR}  # This ensures the targets can find  the headers
     )
-endforeach()
+endforeach ()
 
 # Platform-specific configurations
-if(APPLE)
+if (APPLE)
     # For macOS, use static libraries directly
-    foreach(target catl-validator catl-hasher)
+    foreach (target catl-validator catl-hasher)
         target_link_libraries(${target} PRIVATE
                 ${BOOST_ROOT}/stage/lib/libboost_iostreams.a
                 ${BOOST_ROOT}/stage/lib/libboost_filesystem.a
                 ${OPENSSL_LIBRARIES}
                 ${ZLIB_LIBRARIES}
         )
-    endforeach()
-else()
+    endforeach ()
+else ()
     # For Linux, use dynamic libraries with static linking option
-    foreach(target catl-validator catl-hasher)
+    foreach (target catl-validator catl-hasher)
         target_link_libraries(${target} PRIVATE
                 -static
                 boost_iostreams
                 ${OPENSSL_LIBRARIES}
                 ${ZLIB_LIBRARIES}
         )
-    endforeach()
-endif()
+    endforeach ()
+endif ()
 
 # Add compiler options
-foreach(target catl-validator catl-hasher)
+foreach (target catl-validator catl-hasher)
     target_compile_options(${target} PRIVATE
             -Wall -Wextra -O2 -Wno-deprecated-declarations
     )
-endforeach()
+endforeach ()
